Apply the split minimum per file, not to the scan total
DataFusion's partitioner tests the scan total against
`repartition_file_min_size`, and the 8 MB threshold inherited that. So a
collection of a thousand 1 MB files cleared the line together, and every
one of them was cut into shares that each re-opened it for a fraction of
its rows -- the opposite of what the threshold is for.

A share pays to open one file, so one file is what has to be large enough
to earn it. `interleaved_file_groups` now cuts each file over the size on
its own and passes the rest through whole, so a mixed scan splits its
large files and leaves its small ones alone. It returns `None` when no
file qualifies, which leaves the listing's own spread in place: a
collection of small files still scans one file per partition, as it
always did.

Nothing changes about how a qualifying file is cut. It is still sliced
into `target_partitions * SLICES_PER_PARTITION` byte ranges, still dealt
round-robin, and the opener still resolves each range against the chunk
list.

+++ beacon-db/beacon-file-formats/beacon-arrow-hdf5/src/source.rs
@@ -61,11 +61,11 @@ use crate::cache::Hdf5ReaderCache;
 /// of a chunk list that has to be built first, so the two are not measuring the
 /// same cost.
 ///
-/// The check is on the scan total, not on one file: that is
-/// [`FileGroupPartitioner`](datafusion::datasource::physical_plan::FileGroupPartitioner)'s
-/// own rule, and a collection of small files is worth splitting when a single
-/// small file is not.
-pub const MIN_SPLIT_SIZE: usize = 8 * 1024 * 1024;
+/// The test is on one file, not on the scan total. A share pays to open a file,
+/// so a file is what has to be large enough to earn it: a collection of small
+/// files still scans in parallel, one file per partition, but none of them is
+/// cut into shares that would each re-open it for a fraction of its rows.
+pub const MIN_SPLIT_SIZE: u64 = 8 * 1024 * 1024;
 
 /// DataFusion [`FileSource`] for HDF5 (`.h5`/`.hdf5`) files.
 #[derive(Debug, Clone)]
